Move migration rules to the separate file
