+++ Patches/PersistedDataPatches+Directory.swift
@@ -34,7 +34,7 @@ extension PersistedDataPatch {
         PersistedDataPatch(version: "167.3.0", block: AvailabilityBehaviourChange.notifyAvailabilityBehaviourChange),
         PersistedDataPatch(version: "198.0.0", block: ZMConversation.introduceParticipantRoles),
         PersistedDataPatch(version: "220.0.4", block: InvalidConnectionRemoval.removeInvalid),
-        PersistedDataPatch(version: "234.0.0", block: AppLockController.migrateKeychainItems) //TODO: katerina check the version
+        PersistedDataPatch(version: "234.0.0", block: TransferApplockKeychain.migrateKeychainItems) //TODO: katerina check the version
     ]
 
 }
