+++ Source/Utilis/Keychain/PasscodeKeychainItem.swift
@@ -59,56 +59,3 @@ public enum PasscodeKeychainItem: KeychainItem {
         return query
     }
 }
-
-// MARK: - Migration rules
-
-public extension PasscodeKeychainItem {
-    
-    static func migratePasscode(in moc: NSManagedObjectContext) {
-        let selfUser = ZMUser.selfUser(in: moc)
-        
-        guard let data = try? Keychain.fetchItem(Deprecated_PasscodeKeychainItem.passcode) else {
-            return
-        }
-        
-        do {
-            try Keychain.updateItem(PasscodeKeychainItem.passcode(selfUser.remoteIdentifier), value: data)
-        } catch {
-            fatalError("Failed to migrate passcode: \(error)")
-        }
-    }
-    
-    enum Deprecated_PasscodeKeychainItem: KeychainItem {
-        case passcode
-        
-        public var uniqueIdentifier: String {
-            return "com.wire.passcode"
-        }
-        
-        public var queryForGettingValue: [CFString: Any] {
-            let query: [CFString: Any]
-            
-            switch self {
-            case .passcode:
-                query = [kSecClass: kSecClassGenericPassword,
-                         kSecAttrAccount: uniqueIdentifier,
-                         kSecReturnData: true]
-            }
-            
-            return query
-        }
-        
-        public func queryForSetting(value: Data) -> [CFString: Any] {
-            let query: [CFString: Any]
-            
-            switch self {
-            case .passcode:
-                query = [kSecClass: kSecClassGenericPassword,
-                         kSecAttrAccount: uniqueIdentifier,
-                         kSecValueData: value]
-            }
-            
-            return query
-        }
-    }
-}
