+++ Tests/Source/Model/FeatureConfiguration/AppLock/AppLockControllerTest.swift
@@ -196,45 +196,3 @@ final class AppLockControllerTest: ZMBaseManagedObjectTest {
         XCTAssertNotEqual(sut.config.appLockTimeout, 30)
     }
 }
-
-// MARK: - Migration tests
-
-extension AppLockControllerTest {
-    func testItMigratesIsActiveStateFromTheKeychainToTheMOC() {
-        //given
-        let selfUser = ZMUser.selfUser(in: uiMOC)
-        let config = AppLockController.Config(useBiometricsOrAccountPassword: false,
-                                              useCustomCodeInsteadOfAccountPassword: false,
-                                              forceAppLock: false,
-                                              timeOut: 900)
-        let sut = AppLockController(config: config, selfUser: selfUser)
-        XCTAssertFalse(sut.isActive)
-        
-        //when
-        let data = ("YES").data(using: .utf8)!
-        ZMKeychain.setData(data, forAccount: AppLockController.FeatureName.lockApp.rawValue)
-        
-        AppLockController.migrateIsApplockActiveState(in: uiMOC)
-        
-        //then
-        XCTAssertTrue(sut.isActive)
-    }
-    
-    func testItDoesNotMigrateIsActiveStateFromTheKeychainToTheMOC_IfKeychainIsEmpty() {
-        //given
-        let selfUser = ZMUser.selfUser(in: uiMOC)
-        let config = AppLockController.Config(useBiometricsOrAccountPassword: false,
-                                              useCustomCodeInsteadOfAccountPassword: false,
-                                              forceAppLock: false,
-                                              timeOut: 900)
-        let sut = AppLockController(config: config, selfUser: selfUser)
-        XCTAssertFalse(sut.isActive)
-        
-        //when
-        ZMKeychain.deleteAllKeychainItems(withAccountName: AppLockController.FeatureName.lockApp.rawValue)
-        AppLockController.migrateIsApplockActiveState(in: uiMOC)
-        
-        //then
-        XCTAssertFalse(sut.isActive)
-    }
-}
