--- Tests/Source/Utils/PasscodeKeychainItem+Migration.swift
@@ -1,44 +1,0 @@
-//
-// Wire
-// Copyright (C) 2020 Wire Swiss GmbH
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see http://www.gnu.org/licenses/.
-//
-
-import XCTest
-@testable import WireDataModel
-
-class PasscodeKeychainItemMigration: ModelObjectsTests {
-    
-    override func tearDown() {
-        Keychain.deletePasscode(for: selfUser.remoteIdentifier)
-        
-        super.tearDown()
-    }
-    
-    func testThatItMigratesAPasscode() {
-        // given
-        let passcode = Keychain.fetchPasscode(for: selfUser.remoteIdentifier)
-        XCTAssertNil(passcode)
-        
-        // when
-        let data = ("123456").data(using: .utf8)!
-        try? Keychain.updateItem(PasscodeKeychainItem.Deprecated_PasscodeKeychainItem.passcode, value: data)
-        
-        PasscodeKeychainItem.migratePasscode(in: uiMOC)
-        
-        // then
-        XCTAssertNotNil(Keychain.fetchPasscode(for: selfUser.remoteIdentifier))
-    }
-}
